- fixed server save bug

+++ Utopia/Utopia.Server/Managers/SQLiteStorageManager.cs
@@ -378,7 +378,7 @@ namespace Utopia.Server.Managers
                 if (reader.Read())
                 {
                     if (reader.IsDBNull(0)) return 0;
-                    var maxNumber = reader.GetInt32(0);
+                    var maxNumber = reader.GetInt64(0);
                     return (uint) maxNumber;
                 } return 0;
             }
